# Documentation/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillShare Network",
    "description": "REST API платформы SkillShare Network. Большинство эндпоинтов требуют заголовок `Authorization: Bearer <JWT>`. Получите токен через `POST /api/v1/auth/login`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Мой профиль",
        "description": "Возвращает профиль текущего пользователя. Требуется Bearer JWT.",
        "operationId": "get_me_api_v1_users_me_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Обновить профиль",
        "description": "Изменяет имя и/или URL аватара. Успех — 200 OK.",
        "operationId": "update_me_api_v1_users_me_patch",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Удалить аккаунт",
        "description": "Мягкое удаление: отменяет активные сделки. Успех — 204 No Content.",
        "operationId": "delete_me_api_v1_users_me_delete",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/skills": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Заменить навыки",
        "description": "Полностью заменяет списки offered/wanted. Неизвестные skill_id — 400.",
        "operationId": "update_my_skills_api_v1_users_me_skills_put",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSkillsPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update My Skills Api V1 Users Me Skills Put"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Мои навыки",
        "description": "Списки предлагаемых и желаемых навыков с названиями из справочника.",
        "operationId": "get_my_skills_api_v1_users_me_skills_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get My Skills Api V1 Users Me Skills Get"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Сменить пароль",
        "description": "Успех — 204 No Content. Неверный текущий пароль — 400 Bad Request.",
        "operationId": "change_password_api_v1_users_me_password_patch",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/reviews": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Мои отзывы",
        "description": "Отзывы, полученные текущим пользователем (до 20 последних).",
        "operationId": "get_my_reviews_api_v1_users_me_reviews_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewOut"
                  },
                  "title": "Response Get My Reviews Api V1 Users Me Reviews Get"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден (404 Not Found)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Объявление не найдено"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (403 Forbidden)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Only exchange members have access"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт данных (409 Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пользователь с таким email уже зарегистрирован"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasswordChangeRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password",
            "description": "Текущий пароль"
          },
          "new_password": {
            "type": "string",
            "minLength": 10,
            "title": "New Password",
            "description": "Новый пароль (минимум 10 символов)"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeRequest"
      },
...
      "ReviewOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exchange_id": {
            "type": "integer",
            "title": "Exchange Id"
          },
          "reviewer_id": {
            "type": "integer",
            "title": "Reviewer Id"
          },
          "reviewer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Name"
          },
          "reviewed_id": {
            "type": "integer",
            "title": "Reviewed Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exchange_id",
          "reviewer_id",
          "reviewed_id",
          "rating",
          "comment"
        ],
        "title": "ReviewOut"
      },
...
      "UserProfile": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "rating": {
            "type": "number",
            "title": "Rating",
            "description": "Средний рейтинг по отзывам",
            "examples": [
              4.5
            ]
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "avatar_url",
          "rating",
          "role"
        ],
        "title": "UserProfile",
        "description": "Полный профиль текущего пользователя."
      },
...
      "UserSkillOffered": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id",
            "description": "ID навыка из справочника",
            "examples": [
              1
            ]
          },
          "level": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Level",
            "description": "Уровень владения 1–3",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "level"
        ],
        "title": "UserSkillOffered"
      },
      "UserSkillWanted": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id",
            "description": "ID желаемого навыка",
            "examples": [
              2
            ]
          },
          "desired_level": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Desired Level",
            "description": "Желаемый уровень 1–3",
            "examples": [
              1
            ]
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "desired_level"
        ],
        "title": "UserSkillWanted"
      },
      "UserSkillsPayload": {
        "properties": {
          "offered": {
            "items": {
              "$ref": "#/components/schemas/UserSkillOffered"
            },
            "type": "array",
            "title": "Offered",
            "description": "Чему могу научить"
          },
          "wanted": {
            "items": {
              "$ref": "#/components/schemas/UserSkillWanted"
            },
            "type": "array",
            "title": "Wanted",
            "description": "Чему хочу научиться"
          }
        },
        "type": "object",
        "title": "UserSkillsPayload",
        "description": "Полная замена навыков пользователя."
      },
...
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name",
            "examples": [
              "Новое имя"
            ]
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "examples": [
              "https://cdn.example.com/avatar.png"
            ]
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Поля которые пользователь может изменить сам."
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT из ответа POST /api/v1/auth/login"
      }
    }
  },
...
}